Milestone 1...
Managed to get all the functionality done. All that's left to do now is to port it to the cloud9 IDE for the demo.

There's probably not going to be any more development on the front end or back end. I haven't tested it too extensively, but it atleast seems to work.